feat(storage): enable metric collector in compactor node.
