+++ src/common/src/lib.rs
@@ -46,6 +46,7 @@ pub mod catalog;
 pub mod collection;
 pub mod config;
 pub mod hash;
+pub mod service;
 #[cfg(test)]
 pub mod test_utils;
 pub mod types;
