+++ src/risedevtool/src/config_gen/prometheus_gen.rs
@@ -46,6 +46,14 @@ impl PrometheusGen {
             .map(|node| format!("\"{}:{}\"", node.address, node.port))
             .join(",");
 
+        let compactor_targets = config
+            .provide_compactor
+            .as_ref()
+            .unwrap()
+            .iter()
+            .map(|node| format!("\"{}:{}\"", node.exporter_address, node.exporter_port))
+            .join(",");
+
         format!(
             r#"# --- THIS FILE IS AUTO GENERATED BY RISEDEV ---
 global:
@@ -69,6 +77,10 @@ scrape_configs:
     metrics_path: /minio/v2/metrics/cluster
     static_configs:
     - targets: [{minio_targets}]
+
+  - job_name: compactor-job
+    static_configs:
+      - targets: [{compactor_targets}]
 "#,
         )
     }
