+++ src/risedevtool/src/service_config.rs
@@ -129,6 +129,7 @@ pub struct PrometheusConfig {
     pub provide_compute_node: Option<Vec<ComputeNodeConfig>>,
     pub provide_meta_node: Option<Vec<MetaNodeConfig>>,
     pub provide_minio: Option<Vec<MinioConfig>>,
+    pub provide_compactor: Option<Vec<CompactorConfig>>,
 }
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
